nirvaah/nirvaah-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.7
httpx==0.24.1
twilio==8.10.0
anthropic==0.18.0
supabase==2.3.0
celery==5.3.6
redis==5.0.1
upstash-redis==0.15.0
scikit-learn==1.4.0
xgboost==2.0.3
joblib==1.3.2
google-cloud-vision==3.7.0
gspread==6.0.0
google-auth==2.27.0
pandas==2.2.0
numpy==1.26.4
python-dotenv==1.0.0
